data_processor_app_backend/requirements.txt
Django~=5.1.2
djangorestframework~=3.15.2
django-cors-headers~=4.6.0
pandas~=2.2.3
numpy~=2.1.2
openpyxl~=3.1.5